+++ pyhearts/_morphology/feature/__init__.py
@@ -1,6 +1,14 @@
 #processing functions for ecg_param 
 
-from .HRV import calc_hrv_metrics
+from .HRV import calc_hrv_metrics, rr_intervals_ms_from_r_peaks
 from .intervals import calc_intervals, interval_ms
 from .shape import extract_shape_features
+
+__all__ = [
+    "calc_hrv_metrics",
+    "rr_intervals_ms_from_r_peaks",
+    "calc_intervals",
+    "interval_ms",
+    "extract_shape_features",
+]